Path navigator widget organization, partial dialogs colocation
