+++ ui/app/src/components/Dialogs/DialogFooter.tsx
@@ -15,26 +15,24 @@
  */
 
 import React from 'react';
-import MuiDialogContent from '@material-ui/core/DialogContent';
+import MuiDialogActions from '@material-ui/core/DialogActions';
 import makeStyles from '@material-ui/styles/makeStyles';
 import createStyles from '@material-ui/styles/createStyles';
-import { Theme } from '@material-ui/core';
-import { palette } from '../styles/theme';
-import clsx from 'clsx';
+import { palette } from '../../styles/theme';
 
-const styles = makeStyles((theme: Theme) => createStyles({
-  dialogContent: {
-    padding: theme.spacing(2),
-    backgroundColor: palette.gray.light0,
-    display: 'flex',
-    flexDirection: 'column',
-    flex: '1 1 auto'
-  }
-}));
+const styles = makeStyles(() =>
+  createStyles({
+    dialogActions: {
+      background: palette.white,
+      borderTop: '1px solid rgba(0, 0, 0, 0.12)',
+      minHeight: '55px'
+    }
+  })
+);
 
-export default function DialogBody(props) {
+export default function DialogFooter(props) {
   const classes = styles({});
   return (
-    <MuiDialogContent classes={{ root: clsx(classes.dialogContent, props.classes?.root) }} {...props} />
+    <MuiDialogActions className={classes.dialogActions} classes={props.classes} {...props} />
   );
 }
