+++ ui/app/src/components/Dialogs/DialogHeader.tsx
@@ -17,7 +17,7 @@
 import { makeStyles } from '@material-ui/core/styles';
 import Typography, { TypographyProps } from '@material-ui/core/Typography';
 import IconButton, { IconButtonProps } from '@material-ui/core/IconButton';
-import { palette } from '../styles/theme';
+import { palette } from '../../styles/theme';
 import MuiDialogTitle from '@material-ui/core/DialogTitle/DialogTitle';
 import CloseIconRounded from '@material-ui/icons/CloseRounded';
 import ArrowBack from '@material-ui/icons/ArrowBackIosRounded';
