+++ ui/app/src/components/Navigation/PathNavigator/PathNavigatorBreadcrumbs.tsx
@@ -1,0 +1,123 @@
+/*
+ * Copyright (C) 2007-2020 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as published by
+ * the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { SandboxItem } from '../../../models/Item';
+import { useStyles } from './styles';
+import React, { useState } from 'react';
+import clsx from 'clsx';
+import SearchBar from '../../Controls/SearchBar';
+import MuiBreadcrumbs from '@material-ui/core/Breadcrumbs';
+import NavigateNextIcon from '@material-ui/icons/NavigateNextRounded';
+import Link from '@material-ui/core/Link';
+import Typography from '@material-ui/core/Typography';
+import IconButton from '@material-ui/core/IconButton';
+import MoreVertIcon from '@material-ui/icons/MoreVertRounded';
+import SearchRoundedIcon from '@material-ui/icons/SearchRounded';
+import { isNavigable } from './utils';
+
+interface BreadcrumbsProps {
+  keyword: string;
+  breadcrumb: SandboxItem[];
+  onMenu(element: Element): void;
+  onSearch(keyword: string): void;
+  onCrumbSelected(breadcrumb: SandboxItem): void;
+}
+
+// PathBreadcrumbs + PathOptions + (Path)Search
+export default function (props: BreadcrumbsProps) {
+  const classes = useStyles({});
+  const { breadcrumb, onCrumbSelected, onMenu, keyword, onSearch } = props;
+  const [showSearch, setShowSearch] = useState(false);
+
+  const onChange = (keyword: string) => {
+    if (keyword === '') {
+      setShowSearch(false);
+    }
+    onSearch(keyword);
+  };
+
+  const maxIndex = breadcrumb.length - 1;
+
+  return (
+    <section className={clsx(classes.breadcrumbs, classes.widgetSection)}>
+      {showSearch ? (
+        <SearchBar
+          autoFocus
+          onChange={onChange}
+          keyword={keyword}
+          showActionButton={true}
+          onActionButtonClick={() => onChange('')}
+          classes={{ root: classes.searchRoot }}
+        />
+      ) : (
+        <>
+          <MuiBreadcrumbs
+            maxItems={2}
+            aria-label="Breadcrumbs"
+            separator={<NavigateNextIcon fontSize="small" />}
+            classes={{
+              ol: classes.breadcrumbsList,
+              separator: classes.breadcrumbsSeparator
+            }}
+          >
+            {breadcrumb.map((item: SandboxItem, i: number) =>
+              maxIndex !== i || (maxIndex === i && isNavigable(item)) ? (
+                <Link
+                  key={item.id}
+                  color="inherit"
+                  component="button"
+                  variant="subtitle2"
+                  underline="always"
+                  TypographyClasses={{
+                    root: clsx(
+                      classes.breadcrumbsTypography,
+                      maxIndex === i && classes.breadcrumbLast
+                    )
+                  }}
+                  onClick={() => onCrumbSelected(item)}
+                  children={item.label}
+                />
+              ) : (
+                <Typography
+                  key={item.id}
+                  variant="subtitle2"
+                  className={classes.breadcrumbsTypography}
+                  children={item.label}
+                />
+              )
+            )}
+          </MuiBreadcrumbs>
+          <div className={clsx(classes.optionsWrapper, classes.optionsWrapperOver)}>
+            <IconButton
+              aria-label="options"
+              className={clsx(classes.iconButton)}
+              onClick={(event) => onMenu(event.currentTarget)}
+            >
+              <MoreVertIcon />
+            </IconButton>
+            <IconButton
+              aria-label="search"
+              className={clsx(classes.iconButton)}
+              onClick={() => setShowSearch(true)}
+            >
+              <SearchRoundedIcon />
+            </IconButton>
+          </div>
+        </>
+      )}
+    </section>
+  );
+}
