+++ ui/app/src/modules/Content/Dependencies/DependenciesDialog.tsx
@@ -41,15 +41,15 @@ import MoreVertIcon from '@material-ui/icons/MoreVertRounded';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import Dialog from '@material-ui/core/Dialog';
-import DialogHeader from '../../../components/DialogHeader';
-import DialogBody from '../../../components/DialogBody';
+import DialogHeader from '../../../components/Dialogs/DialogHeader';
+import DialogBody from '../../../components/Dialogs/DialogBody';
 import Chip from '@material-ui/core/Chip';
 import CreateIcon from '@material-ui/icons/CreateRounded';
 import InsertDriveFileOutlinedIcon from '@material-ui/icons/InsertDriveFileOutlined';
 import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 import { SuspenseWithEmptyState } from '../../../components/SystemStatus/Suspencified';
-import DialogFooter from '../../../components/DialogFooter';
+import DialogFooter from '../../../components/Dialogs/DialogFooter';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
 import Radio from '@material-ui/core/Radio';
