+++ ui/app/src/modules/Content/Publish/PublishDialog.tsx
@@ -36,10 +36,10 @@ import DependencySelection from '../Dependencies/DependencySelection';
 import PublishForm from './PublishForm';
 import { withStyles } from '@material-ui/core/styles';
 import Dialog from '@material-ui/core/Dialog';
-import DialogHeader from '../../../components/DialogHeader';
-import DialogBody from '../../../components/DialogBody';
+import DialogHeader from '../../../components/Dialogs/DialogHeader';
+import DialogBody from '../../../components/Dialogs/DialogBody';
 import { SuspenseWithEmptyState } from '../../../components/SystemStatus/Suspencified';
-import DialogFooter from '../../../components/DialogFooter';
+import DialogFooter from '../../../components/Dialogs/DialogFooter';
 import Button from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { palette } from '../../../styles/theme';
