+++ ui/app/src/modules/System/Sites/Create/CreateSiteDialog.tsx
@@ -30,7 +30,7 @@ import BlueprintForm from './BlueprintForm';
 import BlueprintReview from './BlueprintReview';
 import LoadingState from '../../../../components/SystemStatus/LoadingState';
 import ErrorState from '../../../../components/SystemStatus/ErrorState';
-import ConfirmDialog from '../../../../components/UserControl/ConfirmDialog';
+import ConfirmDialog from '../../../../components/Dialogs/ConfirmDialog';
 import { Blueprint } from '../../../../models/Blueprint';
 import { CreateSiteMeta, MarketplaceSite, SiteState, Views } from '../../../../models/Site';
 import { defineMessages, useIntl } from 'react-intl';
@@ -54,9 +54,9 @@ import { fadeIn } from 'react-animations';
 import { Subscription } from 'rxjs';
 import SearchBar from '../../../../components/Controls/SearchBar';
 import { useEnv, useSpreadState } from '../../../../utils/hooks';
-import DialogHeader from '../../../../components/DialogHeader';
-import DialogBody from '../../../../components/DialogBody';
-import DialogFooter from '../../../../components/DialogFooter';
+import DialogHeader from '../../../../components/Dialogs/DialogHeader';
+import DialogBody from '../../../../components/Dialogs/DialogBody';
+import DialogFooter from '../../../../components/Dialogs/DialogFooter';
 import Checkbox from '@material-ui/core/Checkbox';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Typography from '@material-ui/core/Typography';
